test crashing if no job queue
