+++ src/utils/db/server.ts
@@ -15,7 +15,7 @@ import { JobQueueManager } from '../../main/job-queues/job-queue-manager'
 import { PluginsServer, PluginsServerConfig } from '../../types'
 import { EventsProcessor } from '../../worker/ingestion/process-event'
 import { status } from '../status'
-import { createPostgresPool, createRedis, UUIDT } from '../utils'
+import { createPostgresPool, createRedis, killProcess, UUIDT } from '../utils'
 import { DB } from './db'
 import { KafkaProducerWrapper } from './kafka-producer-wrapper'
 
@@ -168,7 +168,16 @@ export async function createServer(
     // :TODO: This is only used on worker threads, not main
     server.eventsProcessor = new EventsProcessor(server as PluginsServer)
     server.jobQueueManager = new JobQueueManager(server as PluginsServer)
-    await server.jobQueueManager.connectProducer()
+
+    try {
+        await server.jobQueueManager.connectProducer()
+    } catch (error) {
+        status.error('🔴', `Can not start job queue producer!`)
+        Sentry.captureException(error)
+        if (server.CRASH_IF_NO_PERSISTENT_JOB_QUEUE) {
+            killProcess()
+        }
+    }
 
     const closeServer = async () => {
         server.mmdbUpdateJob?.cancel()
